optimized A* further renamed some node attributes
The optimization to A* involved replacing several usages of dicts with usage of attributes. This reduced hashing overhead, and resulted in a 15.8% on map 02.

It must be noted that this makes several attributes off-limits for Node implementations. These attributes are:
_h
_g
_came_from

The changes to RectNode were made for precisely this reason.